Update pwm.h
The pwm.h has to include the Owtimer.h first and then Owtask.h. If Owtask.h is include first, then in the pwm.cpp, the compiler will align the mTimeout_ms to 4 bytes, while the owTimer.cpp is 8 bytes. So the final program will crash or run unexpectedly. This is the compiler bug I think. In order to include the Ostask.h first. then I need to pad the Owtimer.h with 7 bytes right before  mTimeout_ms .